requirements.txt
altgraph==0.17.3
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
future==0.18.2
idna==3.4
MouseInfo==0.1.3
mutagen==1.46.0
pefile==2022.5.30
Pillow==9.3.0
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodomex==3.15.0
PyGetWindow==0.0.9
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.13
PyMsgBox==1.0.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyRect==0.2.0
PyScreeze==0.1.28
python-decouple==3.6
pytweening==1.0.4
pywin32-ctypes==0.2.0
requests==2.28.1
speedtest-cli==2.1.3
style==1.1.0
update==0.0.1
urllib3==1.26.12
websockets==10.4
yt-dlp==2022.11.11
